requirements.txt
# Development Tools
build
pytest
wheel

# Documentation Dependencies
markdown-it-py==2.0.1 
myst-parser==0.16.1 
sphinx==4.5.0
sphinx-jsonschema==1.17.2
sphinx-click==4.3.0
sphinx_rtd_theme==1.2.0

# Install clpipe in edit mode
-e .

